Remove type casting for NavigationId in endpoint
Updated the UpdateNavigationStatusEndpoint to remove the casting of `req.NavigationId` to `short`. This change allows the `_navigationService.UpdateNavigation` method to accept `req.NavigationId` in its original type, improving type safety and clarity.

+++ src/services/src/Demoulas.ProfitSharing.Endpoints/Endpoints/Navigations/UpdateNavigationStatusEndpoint.cs
@@ -29,7 +29,7 @@ public class UpdateNavigationStatusEndpoint : Endpoint<UpdateNavigationRequestDt
 
     public override async Task<UpdateNavigationStatusResponseDto> ExecuteAsync(UpdateNavigationRequestDto req, CancellationToken ct)
     {
-    var isSuccessful  = await _navigationService.UpdateNavigation((short)req.NavigationId, req.StatusId, cancellationToken: ct);
+    var isSuccessful  = await _navigationService.UpdateNavigation(req.NavigationId, req.StatusId, cancellationToken: ct);
         var response = new UpdateNavigationStatusResponseDto { IsSuccessful = isSuccessful };
         return response;
     }
